fix(remote): update stale agentboot/ask imports after ask package move
The ask package moved to remote/control/ask but five files (imchannel + bot
prompt_reply, incl. tests) still imported the deleted agentboot/ask path,
breaking the root module build.
